api/requirements.txt
fastapi==0.115.6
uvicorn==0.33.0
celery==5.3.4
redis==5.0.0
python-dotenv==1.0.1
python-multipart == 0.0.19
boto3==1.35.81
pymongo == 4.10.1
requests==2.32.3
numpy==2.1.2
webrtcvad